MS-70: move Confirmation component to Modal folder
