+++ src/Components/Shared/Modal/index.js
@@ -2,7 +2,7 @@ import React from 'react';
 import styles from './modal.module.css';
 import Button from '../Button';
 import PostulantsForm from '../../Postulants/Form';
-import Confirmation from '../../Postulants/Confirmation';
+import Confirmation from './Confirmation';
 
 function Modal({ handleShowModal, modalType, meta, handleSubmit }) {
   let modalComponent;
